Destroy active Vulkan objects on replay exit
If instance or device objects are still live at replay exit, replay will
now destroy them and any objects that may have been created from them.
If no instance or device objects were live at replay exit, but other
Vulkan objects are live, replay will report the total number of Vulkan
objects of each type that were leaked.

This is primarily targeted at trimmed files, which may not have recorded
the API calls to destroy objects.  For Android, leaked resource can be
an issue as the replay Activity starts and stops, but the process does
not always exit, so consecutive replays either leak resources until the
process runs out of memory or start with layers in the state from the
previous replay because a VkInstance object was not destroyed and
layers remained loaded by the process.

Change-Id: I78fada7da98e888f6c62f1311f0a85ff66d8fe9e