Ensure orientation is set correctly for Android
Call the replay window's SetSizePreTransform method at replay of
swapchain creation to ensure that we will set the proper screen
orientation for Android when the swapchain pre-transform specifies a 90
or 270 degree rotation with older capture files that do not include a
ResizeWindowCmd2 command. This will do nothing when the current window
properties match the swapchain creation properties, but will change
screen orientation when the current window pre-transform does not match
the swapchain creation pre-transform.

Change-Id: Ia203e6f3d7f398e4582ab705980ee34ffac72521