Fix acquire/present image index mismatch on replay
Override the values of the image indices provided to vkQueuePresentKHR
to ensure that the match the index returned at image acquire.  This
prevents OUT_OF_DATE errors generated by an acquire/present index
mismatch when replaying on Android.  In the case of an index mismatch,
rendering is still performed to the image that was acquired on capture,
and not the image acquired on replay.  This issue will be addressed by
the swapchain portability change.

Change-Id: I3a55a811e71545404ef37d718607fef15e2265df