Support Android replay with multiple surfaces
Android replay currently has a single activity, providing a single
ANativeWindow with support for creation of a single surface/swaphcain.
This change adds a new command line option for replay to specify the
index for a surface to be used for rendering.  Surface/swapchain
creation will only be performed for the specified surface.  Creation for
other surfaces/swapchains will be skipped.  If swapchain creation is
skipped, images will be created to replace that swapchain's images for
rendering.

Change-Id: I098079daa1480a091aca35b7683ef5db50100b86