Update codegen encode/decode for generic handles
Add code generation support for processing generic handle values
with a uint64_t type and an associated enum value specifying the
actual handle type.  The enum value is used to convert the
uint64_t type to the appropriate handle type, which can then be
unwrapped for capture encoding and mapped to a valid handle for
replay decoding.

Change-Id: I2c791016c15e43c4e71740505692f1ae07e20c9b