Init allocator funcs by api version/extensions
When creating the VulkanResourceAllocator during replay, initialize its
function table based on the API version specified at VkInstance creation
and the extensions enabled at VkInstance and VkDevice creation.  The
logic for the function table initialization had looked like:
if (device_table->BindBufferMemory2 != nullptr)
    functions.bind_buffer_memory2 = device_table->BindBufferMemory2;
else
    functions.bind_buffer_memory2 = device_table->BindBufferMemory2KHR;
Because the function pointers in the replay dispatch tables are never
null, this resulted in no-op behavior for extension or version 1.1
functions that were not enabled.

Change-Id: I25d2683cb2d9e584fc99de6c0a8868c3f80e88c3