Warn when a no-op function is called
The function dispatch tables used for capture and replay never contain
null function pointers.  Instead, dispatch table entries for Vulkan
functions that are not available from the current implementation are set
to a no-op function, which would silently do nothing.  This change
updates the no-op functions to print a warning when they are called.

Change-Id: I2dec244238387fa3b9c50b8932d358a7ba13b9ba