Add proxy capture DLLs
Add the initial implementations for the proxy d3d12.dll and dxgi.dll
DLLs to be used for D3D12 capture, with a d3d12_capture.dll that will
containt the main capture implementation to be imported by both of
the proxy DLLs.  The proxy DLLs currently export function pointers for
the undocumented d3d12.dll and dxgi.dll entry points, instead of
actual functions, which does not work correctly.

Change-Id: I032a7cae72f07eadb5b04c0e01a1044051fefcd9