Prevent cmd.exe from closing on replay exit
When a new cmd.exe window is opened at gfxrecon-replay.exe process start,
which will close at process termination, wait for user input before
exiting the process.  This will keep the cmd.exe window open until a key
is pressed, allowing the user to read replay output.

The WIN32 GetConsoleProcessList API call is used to determine if the
cmd.exe instance was created when gfxrecon-replay.exe started.  This
results in the following scenarios:
- If the user started gfxrecon-replay.exe from an existing cmd.exe
  instance, the process will exit normally.
- If the user did not start gfxrecon-replay.exe from an existing
  cmd.exe instance (e.g. the user dragged and dropped a capture file
  onto the gfxrecon-replay.exe program icon in the file explorer, or
  launched gfxrecon-replay.exe from RenderDoc), the window for the
  cmd.exe instance that was created at gfxrecon-replay.exe process
  start will remain open until closed by the user.

Change-Id: I55fb0d17e05d74682d6c3d02961dc8289eba5aef