Consolidate replay phys dev initialization
Adds a member function to VulkanReplayConsumerBase to initialize
some of the VkInstance information that is stored in the
PhysicalDeviceInfo structure, which will be called by both the
vkEnumeratePhysicalDevices and vkEnumeratePhysicalDeviceGroups
replay overrides.  The other VulkanReplayConusmerBase memeber
functions that had been asserting that the replay_device_info field
in PhysicalDeviceInfo should not be null, will also call this
function to initialize the PhysicalDeviceInfo struct's VkInstance
info when that field is null.

Change-Id: I7efa8e1dd40935b771f2b8d0a0484276a287e670