Save and replay opaque address even when zero
If either the VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT or
VK_BUFFER_USAGE_ACCELERATION_STRUCTURE_STORAGE_BIT_KHR usage bits is
set for buffers, or always for acceleration structures, save and replay
their device addresses. The opaque usage should still be enabled at
replay, and it is valid to create these objects with a device address
that is 0.

Change-Id: I54d04edb055e21f41dc147d3492cf2249fe3e314