Add sanity check for supported extensions
If the extension filtering option is not specified, the replayer will
now check that the requested extensions are available on the current
platform and warn the user when this is not the case.
This saves a lot of time on development platforms as debugging
missing extensions without help from the replayer is time consuming.

Change-Id: Ib861427bc92c90e255d4235ad285409d5eb72c4a