Make gfxrecon-capture respect PATH
Update gfxrecon-capture.py so that it no longer prepends the optional
working directory or './' to the executable name.  The executable is
now launched exactly as specified on the command line. The
shutil.which() function is used to determine the full path to the
executable for purposes of verifying that the specified value is a
valid executable.

Change-Id: I252f77af24325cf46f21b2738f948f6322c66ecf