Basic D3D12 class wrapper generation
Initial wrapper code generation implementation, which generates basic
wrapper classes for DXGI and D3D12 classes.  The generated methods
currently only include calls to the wrapped object with parameters
that are not processed for object wrapping/unwrapping, and which are not
recorded to the capture file.  These items will be added in a future
commit.

Change-Id: I21c62628e8b9dd3ec51d62542044efcecd954a53