Add stack count for wrapped functions
Adds a counter, in TLS, that is incremented by each wrapped function or
method call, which is used to determine if the call is being made
directly by the application or by another wrapped API call.  Object
unwrapping and parameter logging will only be performed for API calls
invoked directly by the application.

Change-Id: I8cf22f8f69cb595564cdc682a3ebd89b4c6fe5f2