Add wrapper destructor
D3D objects can be replaced with wrapper objects because they have
compatible vtables.  To maintain vtable compatibility, the virtual
functions that are not part of the D3D interface cannot be added
to the wrappers, including a virtual destructor.  To compensate for the
lack of a virtual destructor, a 'destructor' function has been added to
the wrapper classes, which is passed to the IUnkown_Wrapper base class
through the wrapper destructor, and used by the Release() implementation
to destroy the object when the reference count reaches zero.

Change-Id: Ifc813f0cdfd0c0dc99847cdb6369d548f052dc6d