Use base functions with capture unwrap codegen
Updates the 'struct unwrappers' code generator to use member functions
from the base class, which determine if a structure has members that
are objects needing to be unwrapped, instead of similar functions that
were implemented directly in the struct unwrappers generator.

This addresses a deficiency with the code that had been implemented in
the struct unwrappers generator, which was not detecting a case where
a struct contained a struct that contained an anonymous union that
contained an object.

Change-Id: Iade7411c62222c7ad83fdddbfcbf2d01c982a6fd