Remove D3D12 condition from util test main
Always specify the test main.cpp in the target_sources macro for
the gfxrecon_util_test target in the util CMakeLists.txt file.
This file isn't D3D12 specific, so can be present for all platforms.
Without this file, the list of files specified to target_sources
will be empty on Linux, which will cause CMake project generation
to fail if tests are enabled.