Add overrides for fence signal/wait replay
Add replay overrides for API calls that perform fence signal and wait
operations.

Overrides were added for the following API calls that perform signal
operations:
- ID3D12Fence::Signal
- ID3D12CommandQueue::Signal
- ID3D12Device3::EnqueueMakeResident

Overrides were added for the following API calls that represent wait
operations:
- ID3D12Fence::SetEventOnCompletion

Two data structures are now stored with each ID3D12Fence object:
- A list of "values" specified for fence signal operations.
- A map of "values" to event handles specified for fence wait
  operations.

For signal operations, the fence's event map will be checked for
an entry keyed by the signal value, with the following results:
- If an entry exists, a wait operation has already been scheduled.  The
  event handle will be waited on and removed from the map.
- If an entry does not exist, the value will be added to the fence's
  pending value list.

For wait operations, the fence's signaled value list will be checked for
a match to the wait operation's value, with the following results:
- If an entry exists, a signal operation for the value has already been
  performed.  The event handle will be waited on and the value will be
  removed from the list.
- If an entry does not exist, the value/event handle pair will be added
  to the map.

Waiting on the event signaled by the ID3D12Device1
SetEventOnMultipleFenceCompletion method is not currently supported as
it requires tracking signal operations on multiple fences with either
an any or all signaled requirement.

Change-Id: I215a54270d893fb64d6fea2c6cfc3b968d24a8c4