Encode API family with meta data type
Store ApiFamilyId in the upper 2 bytes of the uint32_t value that
encodes MetaDataType in the capture file. Rename the corresponding
struct member from meta_data_type to meta_data_id. Change MetaDataType
from uint32_t to uint16_t and provide functions to convert ApiFamilyId
and MetaDataType to and from the encoded uint32_t value. Modify
file_processor so a decoder only processes the meta data if the api
family and meta data type are supported by the decoder.

Change-Id: I385a06fbb10d01d7d823398d7976b55e832d9dca