Add custom D3D12_PIPELINE_STATE_STREAM_DESC encode
Add custom encode and decode routines for the
D3D12_PIPELINE_STATE_STREAM_DESC structure to apply the appropriate
encoding and decoding rules to the subobjects in the
pPipelineStateSubobjectStream member based on subobject type.
Subobjects are stored in an array of bytes that the code generator
cannot determine how to process based on the information available from
the header files.

Change-Id: Ic0127e255ba88c0c2f77d01550c9b9e0d102b669