Skip desc offset calc for start addr
When mapping descriptor addresses for replay, if the captured descriptor
address matches the start address of the descriptor heap, skip the
offset calculation and return the heap's replay start address.  This
helps with the case where a heap was created for a single descriptor and
the application does not query for the descriptor increment, which would
cause descriptor mapping to fail because it expects the increment to be
available for offset calculation.

Change-Id: Ie77e1d8e9ac9f1d879571f25648962dc1d486d09