Release unused objects at replay
When an API call that failed on capture succeeds on replay, call Release
for any objects that the API call created.  Because the object was not
created at capture, a Release call for it will not have been written to
the capture file.  Because the object will have a null ID, due to the
capture failure, it will also not be added to the object table to be
cleaned up on exit, and will effectively be leaked.

Change-Id: Ib3ce5f3e9a603d390b5f6ac6d8e48a27666a8667