Pre-wrap swapchain images
Add custom post call commands to capture for swapchain creation calls,
which will preacquire each swapchain image, wrap the image, store it
with the swapchain, and maintain an internal reference to the wrapper.
The internal reference will be released when the swapchain is destroyed,
ensuring that the image wrappers stay active while the swapchain is
active.

Change-Id: Ia038b228ea7ed5ddadd77053eb49d698d3e65b66