Keep swapchain image replay info active
Prevent swapchain image info structures from being removed from the
object info table while the swapchain is active.  The first time the
image is retrieved, its replay reference count will be incremented and
its info struct will be stored with the swapchain.  The swapchain's
references will be released when the swapchain is destroyed or the
buffers are resized.

Change-Id: I14289a791c628e4192611d820c214be38e0f1224