Use EncodeStructArray2D for ppGeometryDescs
Change the function used to encode the acceleration structures
ppGeometryDescs member from EncodeStructArrayPtr to EncodeStructArray2D.
The EncodeStructArrayPtr function appears to have been added for this
value, but it does not have a corresponding decode function, so it is
also removed by this change.  EncodeStructArray2Ds matching decode
function was already being used to decode this value.

This matches the encoding for the correspoding Vulkan ppGeometries
member of the VkAccelerationStructureBuildGeometryInfoKHR structure.

Change-Id: Idd834e66d53be022ba4dcfbcd6af4e4b20f0f5ad