Check GetBuffer result for SC image pre-acquire
When attempting to pre-acquire swap chain images at capture, check the
GetBuffer() result and skip wrapper creation on failure.  If GetBuffer()
fails with E_NOINTERFACE, log a warning that indicates the API being
captured may not be D3D12, as the DXGI shim can be loaded by D3D11
applications.

Change-Id: Ie96e9102a3eba408204666e8fdf89930e52a8512