Update --sync for outstanding queue waits
For the replay --sync option, which inserts and waits on a fence Signal
operation, the wait operation is enqueued in the pending event queue if
the queue has outstanding Wait operations.  This prevents
single-threaded replay from waiting on a event in a way that would
prevent the outstanding Wait from being signaled.

Change-Id: I8b25c597ee4873fbee013d69671e5cc2fd56b6d5