Compare fence wait/signal values with <=
Change the logic used for determining when a fence calue has been
signaled from == to <= when comparing the target fence value to the
last signaled value.  This aligns with run-time behavior documented
as "A signal operation modifies the fence value and a wait operation
blocks until the fence has reached the requested value or greater".

Change-Id: Ia6e068a1cbfff6ea83ec2cc47a101e763089092b