Fix a race condition and crash when done trimming
When a block is done trimming, the file is closed and GFXR sets a state
variable to indicate no additional API calls should be captured to the
file. However there was no lock around changing the trim state and the
file close, so it was possible for a thread to try and write to file
after it was closed. This is fixed by acquiring the unique state lock
when opening and closing the file.

Change-Id: I2655ec4430b20ee01651597410496e5b8dc1b074