Track descriptor copies for trimming
If CopyDescriptors or CopyDescriptorsSimple is called when tracking
state for trimming, the creation calls for the copied descriptors are
copied to the destination heap. This ensures the destination heap will
contain the correct descriptors regardless of whether the source
descriptor heap is modified or destroyed before writing the trim state.

Change-Id: Ia40b9d9e62eaede0e181ce32657dd175089d3334