Store a copy of the IID instead of a reference
This fixes an issue when capturing the output of GFXR replay aka
recapture. Replay creates a temporary IID/GUID object used to make API
calls. Capture then stored a reference instead of a copy. When the temp
object was destroyed after the API call returned, the IID ref was
invalidated.