Fix deferred operation error
This commit includes the following changes on both capture and replay side:
if vkCreateRayTracingPipelinesKHR is a deferred operation,the lifetime of
the input parameters are extended to cover related
vkDeferredOperationJoinKHR and vkGetDeferredOperationResultKHR until the
return value show the deferred operation be finished. The post handling of
vkCreateRayTracingPipelinesKHR now is also placed after the deferred
operation be finished.

Change-Id: If5b011352703cb9d9db54a887d082199f3f24cb1