Don't wait forever for queues to idle on exit
On exit it is possible for a queue to have a pending wait for a signal
that has not been replayed. Add a timeout to the wait on the queues, and
if the wait fails, don't attempt to release DX12 objects that may still
be in use by the runtime.