scripts: add keyframes.py
The keyframes.py script is used to algorithmically find "interesting"
keyframes for a trace file.  It is used to process traces for use in
testing; the identified keyframes of a trace are used as target frames
for image comparison and trimming operations.

The included .pylintrc file is useful on Linux for "linting" scripts.
This .pylintrc reflects LunarG's Python standards.