issue-515: Picking wrong physical device when multiple are present
This is a minor bug that could trigger in setups with multiple ICDs
present.

The logic in GetMatchingDevice was erroneously querying the physical
device information of the currently selected physical device, instead of
the one currently in its valuation list. As a result it was possible to
make a perfect match between two completely different physical devices.