Fix assert in Dx12ResourceValueMapper
Dx12ResourceValueMapper tracks the copies of resource values to
determine which resource the value originated from. If all the values
of a particular resource were copied from another resource, it can be
removed from the list of resources that needs values mapped. This fixes
an assert in Dx12ResourceValueMapper::CopyResourceValues