Unblock SIGSEGV everytime a memory protection is applied on a region
It is possible that SIGSEGV is blocked in a thread's signal mask. As a
result this hinders page guard mechanism, as any user accesses to
protected memory will not trigger the handler but instead terminate the
process with segmentation fault.

This commit adds the possibility to force-enable SIGSEGV each time a
protection is applied on a shadow memory. This can be done by setting
the GFXRECON_PAGE_GUARD_UNBLOCK_SIGSEGV boolean environment variable to
true. Otherwise, if SIGSEGV is blocked, a warning message is printed.