Add util to track FillMemCmd src for DXR data
The Dx12ResourceValueTracker contains the functionality to track which
FillMemoryCommand wrote the data for each value that is mapped by the
Dx12ResourceValueMapper for DXR or EI playback. This will be used in a
post-process optimization tool to map the data in the source
FillMemoryCommand before it is uploaded to CPU. Once optimized, GFXR
will not need to inject extra GPU commands to map the data. The tracking
will require additional tool changes therefore it is not enabled by this
commit.