Fix GetResourceTiling capture crash
This commit changes the source generator which fixes access violation that
causes the crash issue. The crash happens because of the call
of ID3D12Device::GetResourceTiling. This commit also includes the changes
of similiar issues in other 7 functions.

Change-Id: I3a77f3a5a6efc7050cd4cdb59e7aea4725a8486c

* Edited comments

Edited comments to make it more generic.

Change-Id: Iec77715773633bd945432b573913fef59a9bcd37

* Simplified comments

Removed some comments for code review.

Change-Id: I3427fedd0196461ca408fc7e35c959c05f648ed8

* Fix whitespace formatting issue

Remove extra whitespace to fix the formatting issue.

Change-Id: I40b8044528374609fa6daa3b16f8091e76fb5e22