Fix issue when synchronizing cmd list executions
The event object used to synchronize command list executions is shared
by all command lists executed on the same queue. Previously the event
was reset immediately in each call to ExecuteCommandLists. It was also
set each time the sync signal was received. This caused an issue when
multiple ExecuteCommandLists calls were made on a command queue that had
an outstanding wait, resulting in incorrect synchronization behavior.

This commit changes behavior so that the event is reset and wait on for
each signalled sync value.