Toascii - Make Each Vulkan Call a JSON Object With Nested Index
Fixes #765

Nests name and type too. This is intended to be suitable for rendering
onto a single line as JSON Lines in a following commit.
This leaves the output as many valid JSON documents concatenated,
rendering the file as a whole, not valid JSON.
It is still good for eyeballing.