Track and restore added to state obj dependencies
* Crash when replaying trimmed trace file.

Problem:
The title used AddToStateObject() to grow new SO from existing one.
When trim started, the existing SO was released, the state tracker was
not able to recreate the new SO.

Solution:
The capturer tracked AddToStateObject() and save the existing SO info.
When trim started, recreated the SO from the saved info.

* Changed the structure to tracking grow from StateObject.