Check resource binding alignment when page guard is in use
When page guard mechanism is in use, and the application performs cpu
writes and gpu reads/copies to different buffers/images that are bound
to the same page of mapped memory, data loss can occur which can lead to
rendering artifacts. This can happen when resources are bound to device
memory at offsets that are not multiples of the system page size. To
work around this the "Page Guard Align Buffer Sizes" environment
variable must be set to true (see usage for more information).

This commit will print a warning message when resources are bound at
offsets that are not a multiples of the system's page size.