scripts: add keyframes.py script
This script is based upon the script of the same name that we have been using
internally to process capture files for testing use.  This version has had several
improvements added, to simplify and automate newly identified capture processing
requirements.  All such improvements are controlled by command-line arguments, so
the script should behave in a backward-compatible way if the new command-line
arguments are not used.

The improved keyframes.py script differs in these ways from its predecessor:

- improved documentation
- gfxrecon-convert tool instead of gfxrecon-toascii tool
- support a .keyframesrc configuration file for user-customized command line arguments
- explicitly pass text encoding (to satisfy modern pylint)
- allow creation of a command_counts.json file during capture file processing
- allow creation of stub reference images during capture file processing

I've also added a .pylintrc file.  It's typical for Python development to use
such a file to encapsulate script style preferences.  This fairly standard instance
is intended to be used as a baseline; as the project makes other style decisions, the
.pylintrc can be updated to match.