Flush After Presents
Flush after presents to help avoid capture files with incomplete final
blocks. The last block in a capture is often a present since the
logic to quit is evaluated between a present and the next Vulkan
call that would be made.
There is also substantial opportunity for the application to crash
inside its normal shutdown path after the last present but before
the standard library executes its internal flush, especially during
development, because crashes that don't lose data when the application
is going down anyway are not a priority to fix.
By selctively flushing after presents we can take care of these cases
without the cost of the option to flush after every block is written.
This ensures the flush is run for any Vulkan or DX12 API call that
GFXR handles as a frame delimiter.