Flush After Presents
Flush after presents to avoid capture files with incomplete final
blocks. The last block in a capture is often a present since the
logic to quit is evaluated between a present and the next Vulkan
call that would be made. By selctively flushing after presents
we can take care of these cases without the cost (if any) of the
option to flush after every block is written.