Encode pNext fields without valid extensions as null when encoding
At least one game in the wild passes (partially) uninitialized memory to
vkEnumeratePhysicalDeviceGroups; as a result, gfxrecon crashes immediately
on startup trying to traverse the invalid pNext chain. This prevents using
gfxrecon for driver regression tests, which is one of the acceptable use
cases.

Even though the Vulkan spec requires these to be NULL, such errors tend to
be overlooked because drivers often just ignore pNext fields where values
are not expected.

Even if a spec update adds extension to a struct, it's unlikely that the
extension struct is known to us, so just skip encoding pNext if we don't
expect one.