Fix Multiple Resource Sharing the Same GPUVA Range
The problem
There is a bug if multiple resources with different sizes share the same
GPUVA range. When mapping a given GPUVA to resource, GFXR's GpuVaMap
returns the resource whose starting GPUVA is closest to the given address.
However, the returned resource might not cover the entire range we were
looking for.

The solution
Search the GpuVaMap with size if it's available, and return the resource
whose range is covered by that size.