Removed VkPerformanceCounterResultKHR ToString
The VkPerformanceCounterResultKHR type never appears in the API because
it is refered to by a void*, for example by vkGetQueryPoolResults().
We can therefore never see it and have a need to convert it.
We lack the local context to get the type out of the void pointer
so adding in this currently dead code would be a small part of a
hypothetical PR which achieved that feat.
The extension this is part of has also not seen broad support
(VK_KHR_performance_query).