Replay: Add ignorable extensions
There are some extensions that we may want to outright ignore during
replay.  This is especially true if the extension is present on the
capture, but not the replay.
One such extension is the VK_EXT_tool_info extension.  This extension is
exposed by the capture layer and may not be present on the system after
the capture layer is removed.

Tests:
- Android with capture layer present and then removed on replay
   + Android instance for UI can grab all listed extensions and enable
     for replay always.
   + However, the entrypoints were not used, just the extension name.
   + Replay worked with message about ignoring the extension.
- Linux with capture layer present for VulkanInfo
   + VulkanInfo can use the extension's entrypoints as well with the
     --show-tool-props argument.
   + Captured with and without tool props
   + Both worked, the only message appearing on the replay when using
     tool props was the following messages:

[gfxrecon] INFO - Skipping call down to vkGetPhysicalDeviceToolProperties
because extension not supported by underlying Vulkan implementation.
[gfxrecon] INFO - Replay adjusted the vkGetPhysicalDeviceToolPropertiesEXT
array count: capture count = 1, replay count = 0