NaN and Inf Float Handling for Convert and Export
Support for these was added to allow Convert's output to be parsed
in python as part of the diffing scripts used validate Export.

Suppress NaNs and Infinities in Export:

The JSON library turns these numbers into nulls otherwise.
Informal testing has shown these only occuring for unitialised
fields and unions for which the float contents are not valid.