screenshot: Add support to export PNG format
Add the ability for grabbing PNG formatted screenshots during replay and
D3D12 capture.  The PNG format is using STB headers (manually added
under the external directory) with the additional tweak of using the
Zlib we already pull in to increase the compression ratio versus
the original image.
Results are showing about a 40% size reduction compared with raw BMP
files.