Fix memory type selection for staging buffers
This commit fixes an issue with erroneous memory type selection when
binding buffer memory without memory translation, e.g. in cases of
binding replay staging buffers. Currently, the code retrieves the
capture device's memory types, even though we want to bind the
memory directly from the replay device.

This issue led to various replay issues with mapping memory later on
in the replay, and was particularly disruptive when attempting to
replay a trimmed capture on a machine with a GPU of a different vendor
than the capture system.