Moved Containers out of OverrideQueuePresentKHR
Allocating the backing store for all these containers on every call
to the function and then freeing it on every exit can be avoided
by having them be member variables that are cleared on each call.
At least for the vectors this should change their size but not their
capacity, and thus not hitting the heap allocation functions once
a steady state of reserved capacity is reached.