Introducing GFXR plugins
The idea behind GFXR plugins is that one can extend GFXR’s functionality
by adding new function calls that will be called before and after each
Vulkan API call.

The plugins should be kept in a separate repository and their
implementation should be independent of GFXR code base. GFXR declares an
interface that each plugin needs to implement in order to communicate.
This gives the following advantages:
- The plugin’s functionality can be a black box (proprietary) as only
the compiled binary needs to be shared
- GFXR sources will not get polluted with functionality that perhaps
only a few users might find useful.

This can be done both for capture and replay.