Applied Our Code Style
Under Ubuntu 22.04.1 on a system with clang-format version:
"Ubuntu clang-format version 14.0.0-1ubuntu1", I applied this:

    python3 ./scripts/build.py -a x64 -c release --code-style

This code style enforcement should be a NOP on a fresh checkout.
One file was not updated as doing so mysteriously fails the style
check on the github runner. It remains todo:
`framework/decode/descriptor_update_template_decoder.h`