Using userfaultfd in page guard manager
Introducing an alternative memory tracking mode which utilizes the
userfaultfd mechanism provided by the Linux kernel. This new mode should
provide an alternative for Linux and Android applications that are not
possible to be captured with the page_guard mode because they interfere
with the SIGSEGV handler. The new mode is enabled by seeting
GFXRECON_MEMORY_TRACKING_MODE/ebug.gfxrecon.memory_tracking_mode to
uffd.