Write replay measurements into a JSON file
Results from `--measurement-frame-range` can now be written to a JSON
file using the option `--measurement-file`.
In addition, a manager has been created to easily write to the JSON
structure from anywhere in the code.

This option is necessary to retrieve measurements in android
environments where logcat is not available. It also makes parsing
of the measurements by a script easier and allow more detailed
measurements without flooding the logcat.

Example of JSON output:

{
    "application": {
        "capture_name": "/sdcard/capture.gfxr"
    },
    "frame_ranges": [
        {
            "duration": 2.365642154,
            "end_frame": 200,
            "end_time_monotonic": 15666.74495701,
            "fps": 42.27182028816688,
            "frame_count": 100,
            "frame_durations": [
                17316846,
                26862231,
                24499077,
                24042769,

                ...

                19480077,
                16187384,
                16081154,
                18678846,
                18439692
            ],
            "start_frame": 100,
            "start_time_monotonic": 15664.379314856
        }
    ]
}

Change-Id: Ib189b1e666677f174b0e8193bc780e422133f8e5