Remove unnecessary swapchain functions
Remove unnecessary swapchain functions (which did nothing but
really passthrough to the final func).  This actually increased
debug build performance on Linux for my Talos trace from 19+ seconds
of runtime to 17+ seconds, but I'm not sure how universal that will
be.
Also replace a few more "decode" structures with the Vulkan handle
usage instead for breaking some of the tight bindings.