Modify virtual swapchain behavior
Modify virtual swapchain to create a command buffer per queue and
and per swapchain image.  Then when a vkQueuePresent occurs,
perform the blits in the same queue as the presented queue to
reduce race conditions.
This resolves the issue where a QueuePresent is presented inline
in a queue, but has no wait semaphore information.  Instead,
the QueuePresent relied upon the previous submit (also in the same
queue) to perform the synchronization.

Previously, the virtual swapchain had it's own Queue that it used
which caused a race condition resulting in old images being used.

Fixes: #1122