Samsung C++ code
This pulls in the changes made by Samsung engineers to get the
ball rolling with the "toCpp" transition.  Unfortunately, pulling
the source over rolled it all up into one commit, so in order
to not lose some history (and the credit) for their hard work,
we wanted to list the contributing engineers here:

  Andy Poole - a.poole@samsung.com
  Péter Gál - pgal.usz@partner.samsung.com
  Zsolt Borbély - zsborbely.u-szeged@partner.samsung.com
  Tamás Kéri - tk-szeged@partner.samsung.com
  Adrián Rávai

Special thanks to their hard work.