Trying ndk-build
This commit attempts to boost Android compilation times. To achieve this
cmake is dropped in favor of ndk-build. This is done because with
ndk-build is easier to get ABIs to compile in parallel. Dropping gradle
completely was also considered and use a script in its place to build
and assemble the .apk but the idea was dropped as the time saved in
compilation times was not worth the changes this would require in the
CI.

Changed summarized:
- For Android builds cmake is droped in favor of ndk-build
- By default ABIs built are reduced from "arm64-v8a armeabi-v7a x86
x86_64" down to "arm64-v8a armeabi-v7a"
- Choosing which ABIs to build can be done with a command line parameter
to gradle like this: "./gradlew assembleDebug -Pabis="arm64-v8a"
-