Use Vulkan-Headers to find current Vulkan Version
The Vulkan-Headers CMakelists.txt has the necessary logic to decode the
Vulkan API version from itself. This allows removal of
FindVulkanVersion.cmake which duplicated that logic.

The Android build is left unchanged for the time being due to the NDK
provided CMake being older than the version required by Vulkan-Headers.