Write Replay Measurements Into a JSON File
Marius measurements-file diffed against dev

Branch pushed to PR was merged with dev rather than rebased on top of
it so I diffed that branch against dev to make a commit equivalent to
rebasing and squashing it - Andrew.

Results from `--measurement-frame-range` can now be written to a JSON
file using the option `--measurement-file`.

This option is necessary to retrieve measurements in android
environments where logcat is not available. It also makes parsing
of the measurements by a script easier and allow more detailed
measurements without flooding the logcat.

Example of JSON output:

{
    "application": {
        "capture_name": "/sdcard/capture.gfxr"
    },
    "frame_range":
    {
        "duration": 2.365642154,
        "end_frame": 200,
        "end_time_monotonic": 15666.74495701,
        "fps": 42.27182028816688,
        "frame_count": 100,
        "frame_durations": [
            17316846,
            26862231,
            24499077,
            24042769,

            ...

            19480077,
            16187384,
            16081154,
            18678846,
            18439692
        ],
        "start_frame": 100,
        "start_time_monotonic": 15664.379314856
    }
}

Co-authored-by: Andrew Cox <andrew@lunarg.com>