Update to Vulkan-Headers v1.3.264
Patch XML on fly for VkFrameBoundaryEXT.pTag
Fixes update to Vulkan-Headers v1.3.264.
Allows encode and decode of VkFrameBoundaryEXT.pTag to be generated
correctly using the same code paths as for the two existing
structures with a similar field.
This is a workaround for a bug in the upstream vk.xml:
<https://github.com/KhronosGroup/Vulkan-Docs/pull/2240>.

Co-authored-by: Mike Schuchardt <mikes@lunarg.com>